refactor(checker): split object literal computation handlers
