+++ crates/tsz-checker/src/types/computation/object_literal/computation.rs
@@ -7,451 +7,15 @@ use super::super::object_literal_context::ContextualPropertyPresence;
 use crate::context::speculation::DiagnosticSpeculationSnapshot;
 use crate::context::{PartialObjectLiteralInitializer, TypingRequest};
 use crate::state::CheckerState;
-use crate::symbols_domain::name_text::{
-    is_zero_arg_call_like_expr_in_arena, simple_computed_name_expr_text_in_arena,
-};
 use tsz_parser::parser::NodeIndex;
 use tsz_parser::parser::node::NodeAccess;
 use tsz_parser::parser::syntax_kind_ext;
-use tsz_scanner::SyntaxKind;
 use tsz_solver::computation::ContextualTypeContext;
-use tsz_solver::{PropertyInfo, TypeId, Visibility};
+use tsz_solver::{TypeId, Visibility};
 
-const SPREAD_DISPLAY_ORDER_OFFSET: u32 = 1_000_000;
-const SPREAD_DISPLAY_ORDER_STRIDE: u32 = 10_000;
-
-fn rebase_spread_display_property_order(props: &[PropertyInfo], base: u32) -> Vec<PropertyInfo> {
-    let mut props = props.to_vec();
-    props.sort_by_key(|prop| prop.declaration_order);
-    for (index, prop) in props.iter_mut().enumerate() {
-        prop.declaration_order = base.saturating_add(index as u32);
-    }
-    props
-}
-
-fn remove_synthetic_missing_union_spread_props(member_props: &mut [Vec<PropertyInfo>]) {
-    let mut required_names = rustc_hash::FxHashSet::default();
-    for props in member_props.iter() {
-        for prop in props {
-            if !prop.optional {
-                required_names.insert(prop.name);
-            }
-        }
-    }
-    if required_names.is_empty() {
-        return;
-    }
-
-    // Conditional object literal unions are completed with `p?: undefined`
-    // placeholders for display/type-union balance. In an object spread, that
-    // placeholder means the branch omits `p`; it should not materialize as a
-    // spread property when another branch supplies a required `p`.
-    for props in member_props {
-        props.retain(|prop| {
-            !(prop.optional
-                && prop.type_id == TypeId::UNDEFINED
-                && required_names.contains(&prop.name))
-        });
-    }
-}
+use super::computation_support::SPREAD_DISPLAY_ORDER_OFFSET;
 
 impl<'a> CheckerState<'a> {
-    fn object_literal_property_is_typed_variable_initializer(
-        &self,
-        property_elem_idx: NodeIndex,
-    ) -> bool {
-        let Some(object_idx) = self.ctx.arena.parent_of(property_elem_idx) else {
-            return false;
-        };
-        let Some(parent_idx) = self.ctx.arena.parent_of(object_idx) else {
-            return false;
-        };
-        let Some(parent_node) = self.ctx.arena.get(parent_idx) else {
-            return false;
-        };
-        if parent_node.kind != syntax_kind_ext::VARIABLE_DECLARATION {
-            return false;
-        }
-        self.ctx
-            .arena
-            .get_variable_declaration(parent_node)
-            .is_some_and(|var_decl| {
-                var_decl.initializer == object_idx && var_decl.type_annotation.is_some()
-            })
-    }
-
-    /// Decide whether an object-literal property value's literal type should
-    /// be widened to its primitive. Inside a `satisfies T` operand we apply
-    /// tsc's exact `isLiteralOfContextualType` per-property gate (via
-    /// `contextual_type_allows_literal`) and ignore `preserve_literal_types` —
-    /// the user-written `T` is a finer specification than the wrapping
-    /// generic-call's contextual broadening. Outside, the coarser legacy
-    /// policy (preserve whenever the outer object context is non-permissive)
-    /// is kept and the downstream deep-widen compensates.
-    fn should_widen_object_property_literal(
-        &mut self,
-        value_type: TypeId,
-        property_context_type: Option<TypeId>,
-        had_object_context: bool,
-        has_non_widening_source: bool,
-    ) -> bool {
-        if self.ctx.in_const_assertion || has_non_widening_source {
-            return false;
-        }
-        if self.ctx.in_satisfies_operand {
-            // Unconstrained properties (not covered by the satisfies type) preserve their
-            // literal types — tsc's `isLiteralOfContextualType` returns false for absent
-            // properties.
-            let Some(ctx_type) = property_context_type else {
-                return false;
-            };
-            // Skip the recursive `isLiteralOfContextualType` walk for values the widener
-            // cannot transform anyway — functions, plain objects, `any`, etc.
-            let value_is_widenable =
-                crate::query_boundaries::common::is_literal_type(self.ctx.types, value_type)
-                    || crate::query_boundaries::common::is_union_type(self.ctx.types, value_type);
-            if !value_is_widenable {
-                return false;
-            }
-            return !self.contextual_type_allows_literal(ctx_type, value_type);
-        }
-        if self.ctx.preserve_literal_types {
-            return false;
-        }
-        let property_context_preserves_literal = property_context_type.is_some_and(|ct| {
-            !crate::query_boundaries::type_computation::core::is_literal_permissive_object_context(
-                ct,
-            )
-        });
-        !property_context_preserves_literal && !had_object_context
-    }
-
-    fn spread_source_is_unannotated_object_literal_binding(&self, expression: NodeIndex) -> bool {
-        let expression = self.ctx.arena.skip_parenthesized_and_assertions(expression);
-        let Some(node) = self.ctx.arena.get(expression) else {
-            return false;
-        };
-        if node.kind != SyntaxKind::Identifier as u16 {
-            return false;
-        }
-        let Some(sym_id) = self
-            .ctx
-            .binder
-            .resolve_identifier(self.ctx.arena, expression)
-        else {
-            return false;
-        };
-        let Some(symbol) = self.ctx.binder.get_symbol(sym_id) else {
-            return false;
-        };
-        for decl_idx in symbol.declarations.iter().copied() {
-            let Some(decl_node) = self.ctx.arena.get(decl_idx) else {
-                continue;
-            };
-            let Some(var_decl) = self.ctx.arena.get_variable_declaration(decl_node) else {
-                continue;
-            };
-            if var_decl.type_annotation.is_some() {
-                return false;
-            }
-            let initializer = self
-                .ctx
-                .arena
-                .skip_parenthesized_and_assertions(var_decl.initializer);
-            let Some(init_node) = self.ctx.arena.get(initializer) else {
-                return false;
-            };
-            return init_node.kind == syntax_kind_ext::OBJECT_LITERAL_EXPRESSION;
-        }
-        false
-    }
-
-    /// Merge a single spread-contributed property into the running
-    /// `properties` map.
-    ///
-    /// tsc's spread merge rule is asymmetric on the *later* property's
-    /// optionality:
-    /// - When the later property is **required**, it fully overrides the
-    ///   earlier one (the runtime always sees the later value).
-    /// - When the later property is **optional**, the runtime may skip
-    ///   it, so the earlier contribution still applies. The merged read
-    ///   type is the union of both, the merged write type is the union
-    ///   of both write types, and the merged property is required iff
-    ///   *some* contributor was required. `readonly` is intersected.
-    ///
-    /// The unconditional-override path that this replaces broke the
-    /// optional-later case in
-    /// `compiler/conformance/types/spread/objectSpreadStrictNull.ts`,
-    /// where `{ ...definiteString, ...optionalNumber }` should produce
-    /// `{ sn: string | number }`, not `{ sn?: number }`.
-    fn merge_spread_property(
-        &self,
-        properties: &mut rustc_hash::FxHashMap<tsz_common::interner::Atom, PropertyInfo>,
-        prop: &PropertyInfo,
-    ) {
-        use std::collections::hash_map::Entry;
-        match properties.entry(prop.name) {
-            Entry::Vacant(slot) => {
-                slot.insert(prop.clone());
-            }
-            Entry::Occupied(mut slot) => {
-                let merged =
-                    crate::query_boundaries::type_computation::core::merge_object_spread_property(
-                        self.ctx.types,
-                        self.ctx.exact_optional_property_types(),
-                        Some(slot.get()),
-                        prop,
-                    );
-                slot.insert(merged);
-            }
-        }
-    }
-
-    fn variable_declaration_for_symbol_decl(
-        &self,
-        decl_idx: NodeIndex,
-    ) -> Option<&tsz_parser::parser::node::VariableDeclarationData> {
-        let mut current = decl_idx;
-        for _ in 0..4 {
-            let node = self.ctx.arena.get(current)?;
-            if let Some(var_decl) = self.ctx.arena.get_variable_declaration(node) {
-                return Some(var_decl);
-            }
-            current = self.ctx.arena.get_extended(current)?.parent;
-        }
-        None
-    }
-
-    fn expression_is_type_assertion(&self, expr_idx: NodeIndex) -> bool {
-        let expr_idx = self.ctx.arena.skip_parenthesized(expr_idx);
-        self.ctx.arena.get(expr_idx).is_some_and(|node| {
-            node.kind == syntax_kind_ext::AS_EXPRESSION
-                || node.kind == syntax_kind_ext::TYPE_ASSERTION
-        })
-    }
-
-    /// Returns `true` when the symbol's value type comes from a non-fresh source:
-    /// either an explicit variable type annotation or a const-asserted initializer.
-    fn sym_has_non_widening_declared_value_type(&self, sym_id: tsz_binder::SymbolId) -> bool {
-        let Some(symbol) = self.ctx.binder.get_symbol(sym_id) else {
-            return false;
-        };
-
-        std::iter::once(symbol.value_declaration)
-            .chain(symbol.declarations.iter().copied())
-            .any(|decl_idx| {
-                self.variable_declaration_for_symbol_decl(decl_idx)
-                    .is_some_and(|var_decl| {
-                        self.ctx.arena.get(var_decl.type_annotation).is_some()
-                            || self.expression_is_const_assertion(var_decl.initializer)
-                    })
-            })
-    }
-
-    fn identifier_refers_to_non_widening_declared_value_type(&self, node_idx: NodeIndex) -> bool {
-        self.ctx
-            .binder
-            .resolve_identifier(self.ctx.arena, node_idx)
-            .is_some_and(|sym_id| self.sym_has_non_widening_declared_value_type(sym_id))
-    }
-
-    fn object_literal_property_access_literal_type(
-        &mut self,
-        node_idx: NodeIndex,
-    ) -> Option<TypeId> {
-        let node_idx = self.ctx.arena.skip_parenthesized_and_assertions(node_idx);
-        let node = self.ctx.arena.get(node_idx)?;
-        if node.kind != syntax_kind_ext::PROPERTY_ACCESS_EXPRESSION {
-            return None;
-        }
-        if let Some(literal_type) = self.const_array_to_enum_member_literal_type_query(node_idx) {
-            return Some(literal_type);
-        }
-        let access = self.ctx.arena.get_access_expr(node)?;
-        self.imported_array_to_enum_member_literal_type(access.expression, access.name_or_argument)
-    }
-
-    fn object_literal_variable_initializer_symbol(
-        &self,
-        idx: NodeIndex,
-    ) -> Option<tsz_binder::SymbolId> {
-        let parent_idx = self.ctx.arena.get_extended(idx)?.parent;
-        let parent_node = self.ctx.arena.get(parent_idx)?;
-        if parent_node.kind != syntax_kind_ext::VARIABLE_DECLARATION {
-            return None;
-        }
-        let var_decl = self.ctx.arena.get_variable_declaration(parent_node)?;
-        if var_decl.initializer != idx {
-            return None;
-        }
-        self.ctx
-            .binder
-            .get_node_symbol(var_decl.name)
-            .or_else(|| self.resolve_identifier_symbol_without_tracking(var_decl.name))
-    }
-
-    fn record_partial_object_literal_property(
-        &mut self,
-        stack_index: Option<usize>,
-        prop: &PropertyInfo,
-    ) {
-        let Some(stack_index) = stack_index else {
-            return;
-        };
-        if let Some(active) = self
-            .ctx
-            .object_literal_tracking
-            .partial_initializers
-            .get_mut(stack_index)
-        {
-            active.properties.insert(prop.name, prop.clone());
-        }
-    }
-
-    fn pop_partial_object_literal_initializer(&mut self, stack_index: Option<usize>) {
-        let Some(stack_index) = stack_index else {
-            return;
-        };
-        if stack_index + 1 == self.ctx.object_literal_tracking.partial_initializers.len() {
-            self.ctx.object_literal_tracking.partial_initializers.pop();
-        } else if stack_index < self.ctx.object_literal_tracking.partial_initializers.len() {
-            self.ctx
-                .object_literal_tracking
-                .partial_initializers
-                .remove(stack_index);
-        }
-    }
-
-    fn pop_object_literal_contexts(
-        &mut self,
-        marker_this_type: Option<TypeId>,
-        partial_initializer_stack_index: Option<usize>,
-    ) {
-        if marker_this_type.is_some() {
-            self.ctx.this_type_stack.pop();
-        }
-        self.pop_partial_object_literal_initializer(partial_initializer_stack_index);
-    }
-
-    fn contextual_this_type_from_marker(&self, ctx_type: TypeId) -> Option<TypeId> {
-        use crate::query_boundaries::common::ContextualTypeContext;
-
-        let env = self.ctx.type_env.borrow();
-        let ctx_helper = ContextualTypeContext::with_expected_and_options(
-            self.ctx.types,
-            ctx_type,
-            self.ctx.compiler_options.no_implicit_any,
-        );
-        if let Some(this_type) = ctx_helper.get_this_type_from_marker_with_resolver(&*env) {
-            return Some(this_type);
-        }
-
-        let def_id = self.ctx.definition_store.find_def_for_type(ctx_type)?;
-        let body = self.ctx.definition_store.get_body(def_id)?;
-        if body == ctx_type {
-            return None;
-        }
-
-        let body_helper = ContextualTypeContext::with_expected_and_options(
-            self.ctx.types,
-            body,
-            self.ctx.compiler_options.no_implicit_any,
-        );
-        body_helper.get_this_type_from_marker_with_resolver(&*env)
-    }
-
-    fn function_like_has_explicit_signature_annotations(&self, expr_idx: NodeIndex) -> bool {
-        let Some(expr_node) = self.ctx.arena.get(expr_idx) else {
-            return false;
-        };
-        match expr_node.kind {
-            k if k == syntax_kind_ext::ARROW_FUNCTION
-                || k == syntax_kind_ext::FUNCTION_EXPRESSION =>
-            {
-                let Some(func) = self.ctx.arena.get_function(expr_node) else {
-                    return false;
-                };
-                if func.type_annotation.is_some() {
-                    return true;
-                }
-                func.parameters.nodes.iter().any(|&param_node| {
-                    self.ctx
-                        .arena
-                        .get(param_node)
-                        .and_then(|node| self.ctx.arena.get_parameter(node))
-                        .is_some_and(|param| param.type_annotation.is_some())
-                })
-            }
-            k if k == syntax_kind_ext::PARENTHESIZED_EXPRESSION => self
-                .ctx
-                .arena
-                .get_parenthesized(expr_node)
-                .is_some_and(|paren| {
-                    self.function_like_has_explicit_signature_annotations(paren.expression)
-                }),
-            _ => false,
-        }
-    }
-
-    fn simple_computed_name_expr_text_for_duplicates(&self, expr_idx: NodeIndex) -> Option<String> {
-        simple_computed_name_expr_text_in_arena(self.ctx.arena, expr_idx)
-    }
-
-    fn is_zero_arg_call_like_expr_for_duplicates(&self, expr_idx: NodeIndex) -> bool {
-        is_zero_arg_call_like_expr_in_arena(self.ctx.arena, expr_idx)
-    }
-
-    fn simple_computed_call_name_for_duplicates(&self, name_idx: NodeIndex) -> Option<String> {
-        let name_node = self.ctx.arena.get(name_idx)?;
-        if name_node.kind != syntax_kind_ext::COMPUTED_PROPERTY_NAME {
-            return None;
-        }
-        let computed = self.ctx.arena.get_computed_property(name_node)?;
-        if !self.is_zero_arg_call_like_expr_for_duplicates(computed.expression) {
-            return None;
-        }
-        let expr_text = self.simple_computed_name_expr_text_for_duplicates(computed.expression)?;
-        Some(format!("[{expr_text}]"))
-    }
-
-    fn this_options_property_access_receiver(&self, expr_idx: NodeIndex) -> Option<NodeIndex> {
-        let expr_idx = self.ctx.arena.skip_parenthesized_and_assertions(expr_idx);
-        let expr_node = self.ctx.arena.get(expr_idx)?;
-        if expr_node.kind != syntax_kind_ext::PROPERTY_ACCESS_EXPRESSION {
-            return None;
-        }
-        let access = self.ctx.arena.get_access_expr(expr_node)?;
-
-        let base_idx = self
-            .ctx
-            .arena
-            .skip_parenthesized_and_assertions(access.expression);
-        let base_node = self.ctx.arena.get(base_idx)?;
-        if base_node.kind != syntax_kind_ext::PROPERTY_ACCESS_EXPRESSION {
-            return None;
-        }
-        let base_access = self.ctx.arena.get_access_expr(base_node)?;
-        let base_name_is_options = self
-            .ctx
-            .arena
-            .get(base_access.name_or_argument)
-            .and_then(|node| self.ctx.arena.get_identifier(node))
-            .is_some_and(|ident| ident.escaped_text == "options");
-        if !base_name_is_options {
-            return None;
-        }
-
-        let receiver_idx = self
-            .ctx
-            .arena
-            .skip_parenthesized_and_assertions(base_access.expression);
-        self.ctx
-            .arena
-            .get(receiver_idx)
-            .is_some_and(|node| node.kind == SyntaxKind::ThisKeyword as u16)
-            .then_some(receiver_idx)
-    }
-
     pub(crate) fn get_type_of_object_literal_with_request(
         &mut self,
         idx: NodeIndex,
@@ -2171,868 +1735,49 @@ impl<'a> CheckerState<'a> {
                 }
             }
             // Accessor: { get foo() {} } or { set foo(v) {} }
-            else if let Some(accessor) = self.ctx.arena.get_accessor(elem_node) {
-                // Always type-check computed property name expressions for accessors,
-                // even when the identifier can be resolved as a literal name.
-                // E.g., `{ get [e]() {} }` needs TS2304 for undeclared `e`.
-                // We call get_type_of_node directly (not check_computed_property_name)
-                // to avoid triggering TS2467 for type parameters in nested object literals.
-                if let Some(prop_name_node) = self.ctx.arena.get(accessor.name)
-                    && prop_name_node.kind
-                        == tsz_parser::parser::syntax_kind_ext::COMPUTED_PROPERTY_NAME
-                    && let Some(computed) = self.ctx.arena.get_computed_property(prop_name_node)
-                {
-                    self.get_type_of_node(computed.expression);
-                }
-                // Missing body for accessors in object literals is a grammar error.
-                // tsc does NOT emit TS1005 here; it defers to TS2378/TS1049
-                // ("A 'get' accessor must have a body"). We skip TS1005 to avoid
-                // false positives that incorrectly suppress TS5107 deprecation
-                // warnings in the driver's grammar-error priority logic.
-
-                // For setters, check implicit any on parameters (error 7006) and on
-                // the property name itself (error 7032).
-                // When a paired getter exists, the setter parameter type is inferred
-                // from the getter return type (contextually typed, suppress TS7006/7032).
-                if elem_node.kind == syntax_kind_ext::SET_ACCESSOR {
-                    let name_opt = self.get_property_name(accessor.name).or_else(|| {
-                        let prop_name_type = self.get_type_of_node(accessor.name);
-                        crate::query_boundaries::type_computation::access::literal_property_name(
-                            self.ctx.types,
-                            prop_name_type,
-                        )
-                        .map(|atom| self.ctx.types.resolve_atom(atom))
-                    });
-                    let has_paired_getter = name_opt
-                        .as_ref()
-                        .is_some_and(|name| obj_getter_names.contains(name));
-                    // Check if accessor JSDoc has @param type annotations
-                    let accessor_jsdoc = self.get_jsdoc_for_function(elem_idx);
-                    let mut first_param_lacks_annotation = false;
-                    for (pi, &param_idx) in accessor.parameters.nodes.iter().enumerate() {
-                        if let Some(param_node) = self.ctx.arena.get(param_idx)
-                            && let Some(param) = self.ctx.arena.get_parameter(param_node)
-                        {
-                            let has_jsdoc = has_paired_getter
-                                || self.param_has_inline_jsdoc_type(param_idx)
-                                || if let Some(ref jsdoc) = accessor_jsdoc {
-                                    let pname = self.parameter_name_for_error(param.name);
-                                    Self::jsdoc_has_param_type(jsdoc, &pname)
-                                } else {
-                                    false
-                                };
-                            if param.type_annotation.is_none() && !has_jsdoc {
-                                first_param_lacks_annotation = true;
-                            }
-                            self.maybe_report_implicit_any_parameter(param, has_jsdoc, pi);
-                        }
-                    }
-                    // TS7032: emit on property name when the setter has no parameter type
-                    // annotation and no paired getter (TSC checks this at accessor symbol
-                    // resolution time; we emit it here during object literal checking).
-                    if first_param_lacks_annotation
-                        && !has_paired_getter
-                        && self.ctx.no_implicit_any()
-                        && let Some(prop_name) = name_opt.as_deref()
-                    {
-                        use crate::diagnostics::diagnostic_codes;
-                        self.error_at_node_msg(
-                            accessor.name,
-                            diagnostic_codes::PROPERTY_IMPLICITLY_HAS_TYPE_ANY_BECAUSE_ITS_SET_ACCESSOR_LACKS_A_PARAMETER_TYPE,
-                            &[prop_name],
-                        );
-                    }
-                }
-
-                let name_opt = if self.object_literal_computed_key_is_wide_symbol(accessor.name) {
-                    None
-                } else {
-                    self.get_property_name_resolved(accessor.name)
-                };
-                if let Some(name) = name_opt.clone() {
-                    // For non-contextual object literals, TypeScript treats `this` inside
-                    // accessors as the object literal under construction. Provide a
-                    // lightweight synthetic receiver so property access checks (TS2339)
-                    // run during accessor body checking.
-                    let mut pushed_synthetic_this = false;
-                    if marker_this_type.is_none() {
-                        let mut this_props: Vec<PropertyInfo> =
-                            properties.values().cloned().collect();
-                        let name_atom = self.ctx.types.intern_string(&name);
-                        if !this_props.iter().any(|p| p.name == name_atom) {
-                            // Getter-only accessors are readonly in the object type
-                            let is_getter_only = elem_node.kind == syntax_kind_ext::GET_ACCESSOR
-                                && !setter_names.contains(&name_atom);
-                            this_props.push(PropertyInfo {
-                                name: name_atom,
-                                type_id: TypeId::ANY,
-                                write_type: TypeId::ANY,
-                                optional: false,
-                                readonly: is_getter_only,
-                                is_method: false,
-                                is_class_prototype: false,
-                                visibility: Visibility::Public,
-                                parent_id: None,
-                                declaration_order: 0,
-                                is_string_named: false,
-                                is_symbol_named: false,
-                                single_quoted_name: false,
-                            });
-                        }
-                        self.ctx
-                            .this_type_stack
-                            .push(self.ctx.types.factory().object(this_props));
-                        pushed_synthetic_this = true;
-                    }
-
-                    // For getter, infer return type; for setter, use the parameter type
-                    let accessor_type = if elem_node.kind == syntax_kind_ext::GET_ACCESSOR {
-                        // Check getter body/parameters via function checking, but object
-                        // property read type is the getter's return type (not a function type).
-                        self.get_type_of_function(elem_idx);
-                        if accessor.type_annotation.is_none() {
-                            // When a contextual type exists (e.g., `T extends { [k: string]: Types }`),
-                            // pass the contextual property type as the return context so that
-                            // literal types from `as const` are preserved instead of widened.
-                            // Without this, `get x() { return 'boolean' as const }` widens
-                            // to `string` because infer_getter_return_type passes None.
-                            let return_context = contextual_type.and_then(|ctx| {
-                                self.contextual_object_property_type_for_lookup(ctx, &name)
-                            });
-                            // Clear `preserve_literal_types` around the body walk so the
-                            // getter's literal-widening decision is independent of the
-                            // outer `return_expression_type` scope. When the enclosing
-                            // expression is itself a function's return statement
-                            // (`function f() { return { get x() { return 1 } } }`), the
-                            // outer scope sets the flag to preserve the obj literal's
-                            // property literals, but the getter body must make its own
-                            // widening decision — otherwise `readonly x: 1` leaks out
-                            // where tsc emits `readonly x: number`. Mirrors the
-                            // nested-function clearing already in `return_expression_type`.
-                            let prev_preserve_literals = self.ctx.preserve_literal_types;
-                            self.ctx.preserve_literal_types = false;
-                            let result = self.infer_return_type_from_body(
-                                tsz_parser::parser::NodeIndex::NONE,
-                                accessor.body,
-                                return_context,
-                            );
-                            self.ctx.preserve_literal_types = prev_preserve_literals;
-                            result
-                        } else {
-                            self.get_type_from_type_node(accessor.type_annotation)
-                        }
-                    } else {
-                        // Setter: type-check the function body to track variable usage
-                        // (especially for noUnusedParameters/noUnusedLocals checking),
-                        // but use the parameter type annotation for the property type
-                        self.get_type_of_function(elem_idx);
-
-                        // Extract setter write type from first parameter.
-                        // When no type annotation, fall back to the paired getter's
-                        // return type (mirroring tsc's inference behavior).
-                        accessor
-                            .parameters
-                            .nodes
-                            .first()
-                            .and_then(|&param_idx| {
-                                let param = self.ctx.arena.get_parameter_at(param_idx)?;
-                                if param.type_annotation.is_none() {
-                                    None
-                                } else {
-                                    Some(self.get_type_from_type_node(param.type_annotation))
-                                }
-                            })
-                            .or_else(|| {
-                                // No annotation — infer from paired getter's type
-                                let setter_name = name_opt.clone()?;
-                                let name_atom = self.ctx.types.intern_string(&setter_name);
-                                properties.get(&name_atom).map(|p| p.type_id)
-                            })
-                            .unwrap_or(TypeId::ANY)
-                    };
-
-                    if pushed_synthetic_this {
-                        self.ctx.this_type_stack.pop();
-                    }
-
-                    if elem_node.kind == syntax_kind_ext::GET_ACCESSOR {
-                        if accessor.type_annotation.is_none() {
-                            use crate::diagnostics::diagnostic_codes;
-                            let self_refs =
-                                self.collect_property_name_references(accessor.body, &name);
-                            if !self_refs.is_empty() {
-                                self.error_at_node_msg(
-                                    accessor.name,
-                                    diagnostic_codes::IMPLICITLY_HAS_RETURN_TYPE_ANY_BECAUSE_IT_DOES_NOT_HAVE_A_RETURN_TYPE_ANNOTATION,
-                                    &[&name],
-                                );
-                            }
-                        }
-
-                        self.maybe_report_implicit_any_return(
-                            Some(name.clone()),
-                            Some(accessor.name),
-                            accessor_type,
-                            accessor.type_annotation.is_some(),
-                            false,
-                            elem_idx,
-                        );
-                    }
-
-                    // TS2378: A 'get' accessor must return a value.
-                    // Check if the getter has a body but no return statement with a value.
-                    if elem_node.kind == syntax_kind_ext::GET_ACCESSOR && accessor.body.is_some() {
-                        let has_return = self.body_has_return_with_value(accessor.body);
-                        let falls_through = self.function_body_falls_through(accessor.body);
-
-                        if !has_return && falls_through {
-                            use crate::diagnostics::diagnostic_codes;
-                            self.error_at_node(
-                                accessor.name,
-                                "A 'get' accessor must return a value.",
-                                diagnostic_codes::A_GET_ACCESSOR_MUST_RETURN_A_VALUE,
-                            );
-                        }
-                    }
-                    let name_atom = self.ctx.types.intern_string(&name);
-
-                    // Check for duplicate property - but allow getter+setter pairs
-                    // A getter and setter with the same name is valid, not a duplicate
-                    let is_getter = elem_node.kind == syntax_kind_ext::GET_ACCESSOR;
-                    let is_complementary_pair = if is_getter {
-                        setter_names.contains(&name_atom) && !getter_names.contains(&name_atom)
-                    } else {
-                        getter_names.contains(&name_atom) && !setter_names.contains(&name_atom)
-                    };
-                    // Duplicate properties are an error in object literals.
-                    // TS1118 for duplicate get/set accessors, TS1117 for other duplicates.
-                    // Skip for computed property names — tsc only checks static names.
-                    if !skip_duplicate_check
-                        && explicit_property_names.contains(&name_atom)
-                        && !is_complementary_pair
-                        && !self.ctx.has_parse_errors
-                        && (!self.is_js_file() || self.ctx.js_strict_mode_diagnostics_enabled())
-                    {
-                        let is_duplicate_accessor = (is_getter
-                            && getter_names.contains(&name_atom))
-                            || (!is_getter && setter_names.contains(&name_atom));
-                        if is_duplicate_accessor {
-                            self.error_at_node(
-                                accessor.name,
-                                "An object literal cannot have multiple get/set accessors with the same name.",
-                                diagnostic_codes::AN_OBJECT_LITERAL_CANNOT_HAVE_MULTIPLE_GET_SET_ACCESSORS_WITH_THE_SAME_NAME,
-                            );
-                        } else {
-                            let message = format_message(
-                                diagnostic_messages::AN_OBJECT_LITERAL_CANNOT_HAVE_MULTIPLE_PROPERTIES_WITH_THE_SAME_NAME,
-                                &[&name],
-                            );
-                            self.error_at_node(
-                                accessor.name,
-                                &message,
-                                diagnostic_codes::AN_OBJECT_LITERAL_CANNOT_HAVE_MULTIPLE_PROPERTIES_WITH_THE_SAME_NAME,
-                            );
-                        }
-                    }
-                    explicit_property_names.insert(name_atom);
-
-                    if is_getter {
-                        getter_names.insert(name_atom);
-                    } else {
-                        setter_names.insert(name_atom);
-                    }
-
-                    let (acc_str_named, acc_sym_named, acc_single_quoted) =
-                        self.object_literal_member_naming_flags(accessor.name);
-                    // Merge getter/setter into a single property with separate
-                    // read (type_id) and write (write_type) types.
-                    if let Some(existing) = properties.get(&name_atom) {
-                        let existing_order = existing.declaration_order;
-                        let (read_type, write_type) = if is_getter {
-                            // Getter arriving after setter
-                            (accessor_type, existing.write_type)
-                        } else {
-                            // Setter arriving after getter
-                            (existing.type_id, accessor_type)
-                        };
-                        // Both getter and setter exist → not readonly
-                        let prop_info = PropertyInfo {
-                            name: name_atom,
-                            type_id: read_type,
-                            write_type,
-                            optional: false,
-                            readonly: false,
-                            is_method: false,
-                            is_class_prototype: false,
-                            visibility: Visibility::Public,
-                            parent_id: None,
-                            declaration_order: existing_order,
-                            is_string_named: acc_str_named || existing.is_string_named,
-                            is_symbol_named: acc_sym_named || existing.is_symbol_named,
-                            single_quoted_name: acc_single_quoted || existing.single_quoted_name,
-                        };
-                        properties.insert(name_atom, prop_info.clone());
-                        self.record_partial_object_literal_property(
-                            partial_initializer_stack_index,
-                            &prop_info,
-                        );
-                    } else {
-                        // Single accessor so far: getter-only is readonly.
-                        // Set-only: read type is `undefined`.
-                        let readonly = is_getter;
-                        let (read_type, write_type) = if is_getter {
-                            (accessor_type, accessor_type)
-                        } else {
-                            (TypeId::UNDEFINED, accessor_type)
-                        };
-                        let order = prop_order;
-                        prop_order += 1;
-                        let prop_info = PropertyInfo {
-                            name: name_atom,
-                            type_id: read_type,
-                            write_type,
-                            optional: false,
-                            readonly,
-                            is_method: false,
-                            is_class_prototype: false,
-                            visibility: Visibility::Public,
-                            parent_id: None,
-                            declaration_order: order,
-                            is_string_named: acc_str_named,
-                            is_symbol_named: acc_sym_named,
-                            single_quoted_name: acc_single_quoted,
-                        };
-                        properties.insert(name_atom, prop_info.clone());
-                        self.record_partial_object_literal_property(
-                            partial_initializer_stack_index,
-                            &prop_info,
-                        );
-                    }
-                } else {
-                    // Computed accessor name - still type-check the expression and body
-                    self.check_computed_property_name(accessor.name);
-
-                    let mut prop_name_type = TypeId::ANY;
-                    if let Some(prop_name_node) = self.ctx.arena.get(accessor.name)
-                        && prop_name_node.kind
-                            == tsz_parser::parser::syntax_kind_ext::COMPUTED_PROPERTY_NAME
-                        && let Some(computed) = self.ctx.arena.get_computed_property(prop_name_node)
-                    {
-                        prop_name_type = self.get_type_of_node(computed.expression);
-                        if let Some(atom) =
-                            crate::query_boundaries::type_computation::access::literal_property_name(
-                                self.ctx.types,
-                                prop_name_type,
-                            )
-                        {
-                            let is_getter =
-                                elem_node.kind == tsz_parser::parser::syntax_kind_ext::GET_ACCESSOR;
-                            let is_complementary_pair = if is_getter {
-                                setter_names.contains(&atom) && !getter_names.contains(&atom)
-                            } else {
-                                getter_names.contains(&atom) && !setter_names.contains(&atom)
-                            };
-                            if !skip_duplicate_check
-                                && explicit_property_names.contains(&atom)
-                                && !is_complementary_pair
-                                && !self.ctx.has_parse_errors
-                                && (!self.is_js_file()
-                                    || self.ctx.js_strict_mode_diagnostics_enabled())
-                            {
-                                let name = self.ctx.types.resolve_atom(atom).to_string();
-                                use crate::diagnostics::{diagnostic_codes, diagnostic_messages};
-                                let message = crate::diagnostics::format_message(
-                                            diagnostic_messages::AN_OBJECT_LITERAL_CANNOT_HAVE_MULTIPLE_PROPERTIES_WITH_THE_SAME_NAME,
-                                            &[&name],
-                                        );
-                                self.error_at_node(
-                                            accessor.name,
-                                            &message,
-                                            diagnostic_codes::AN_OBJECT_LITERAL_CANNOT_HAVE_MULTIPLE_PROPERTIES_WITH_THE_SAME_NAME,
-                                        );
-                            }
-                            explicit_property_names.insert(atom);
-                        }
-                    }
-
-                    let accessor_type = if elem_node.kind == syntax_kind_ext::GET_ACCESSOR {
-                        self.get_type_of_function(elem_idx);
-
-                        // TS2378: A 'get' accessor must return a value.
-                        if accessor.body.is_some() {
-                            let has_return = self.body_has_return_with_value(accessor.body);
-                            let falls_through = self.function_body_falls_through(accessor.body);
-                            if !has_return && falls_through {
-                                use crate::diagnostics::diagnostic_codes;
-                                self.error_at_node(
-                                    accessor.name,
-                                    "A 'get' accessor must return a value.",
-                                    diagnostic_codes::A_GET_ACCESSOR_MUST_RETURN_A_VALUE,
-                                );
-                            }
-                        }
-
-                        if accessor.type_annotation.is_none() {
-                            self.infer_getter_return_type(accessor.body)
-                        } else {
-                            self.get_type_from_type_node(accessor.type_annotation)
-                        }
-                    } else {
-                        self.get_type_of_function(elem_idx);
-                        accessor
-                            .parameters
-                            .nodes
-                            .first()
-                            .and_then(|&param_idx| {
-                                let param = self.ctx.arena.get_parameter_at(param_idx)?;
-                                if param.type_annotation.is_none() {
-                                    None
-                                } else {
-                                    Some(self.get_type_from_type_node(param.type_annotation))
-                                }
-                            })
-                            .unwrap_or(TypeId::ANY)
-                    };
-
-                    self.route_computed_member_value_to_index_signature(
-                        prop_name_type,
-                        accessor_type,
-                        &mut number_index_types,
-                        &mut string_index_types,
-                        &mut symbol_index_types,
-                    );
-                }
+            else if self.process_object_literal_accessor_element(
+                elem_idx,
+                &obj_getter_names,
+                contextual_type,
+                marker_this_type,
+                &mut properties,
+                &mut setter_names,
+                &mut getter_names,
+                &mut explicit_property_names,
+                skip_duplicate_check,
+                &mut prop_order,
+                partial_initializer_stack_index,
+                &mut number_index_types,
+                &mut string_index_types,
+                &mut symbol_index_types,
+            ) {
             }
             // Spread assignment: { ...obj }
             else if elem_node.kind == syntax_kind_ext::SPREAD_ELEMENT
                 || elem_node.kind == syntax_kind_ext::SPREAD_ASSIGNMENT
             {
-                has_spread = true;
-                let spread_expr = self
-                    .ctx
-                    .arena
-                    .get_spread(elem_node)
-                    .map(|spread| spread.expression)
-                    .or_else(|| {
-                        self.ctx
-                            .arena
-                            .get_unary_expr_ex(elem_node)
-                            .map(|unary| unary.expression)
-                    });
-                if let Some(spread_expr) = spread_expr {
-                    let mut invalid_rest_target = false;
-                    if self.ctx.in_destructuring_target {
-                        // TS2701: The target of an object rest assignment must be
-                        // a variable or a property access.
-                        // E.g. `{ ...expr + expr } = source` is invalid.
-                        if !self.is_valid_rest_assignment_target(spread_expr) {
-                            use crate::diagnostics::{diagnostic_codes, diagnostic_messages};
-                            self.error_at_node(
-                                spread_expr,
-                                diagnostic_messages::THE_TARGET_OF_AN_OBJECT_REST_ASSIGNMENT_MUST_BE_A_VARIABLE_OR_A_PROPERTY_ACCESS,
-                                diagnostic_codes::THE_TARGET_OF_AN_OBJECT_REST_ASSIGNMENT_MUST_BE_A_VARIABLE_OR_A_PROPERTY_ACCESS,
-                            );
-                            invalid_rest_target = true;
-                        }
-                        // TS2778: The target of an object rest assignment may not be
-                        // an optional property access. E.g. `{ ...obj?.a } = source`
-                        else if self.is_optional_chain_access(spread_expr) {
-                            use crate::diagnostics::{diagnostic_codes, diagnostic_messages};
-                            self.error_at_node(
-                                spread_expr,
-                                diagnostic_messages::THE_TARGET_OF_AN_OBJECT_REST_ASSIGNMENT_MAY_NOT_BE_AN_OPTIONAL_PROPERTY_ACCESS,
-                                diagnostic_codes::THE_TARGET_OF_AN_OBJECT_REST_ASSIGNMENT_MAY_NOT_BE_AN_OPTIONAL_PROPERTY_ACCESS,
-                            );
-                        }
-                    }
-                    // Clear contextual type for call-like spread expressions.
-                    // The outer contextual type (e.g., from a destructuring pattern)
-                    // should not propagate into call expression return types —
-                    // otherwise IIFEs in spreads get false contextual return types,
-                    // producing spurious TS2741/TS2322 errors.
-                    // But direct object literals in spreads (e.g., `{ ...{ a: "a" } }`)
-                    // SHOULD keep the contextual type so literals stay narrow.
-                    let unwrapped_spread = self
-                        .ctx
-                        .arena
-                        .skip_parenthesized_and_assertions(spread_expr);
-                    let spread_is_call_like =
-                        self.ctx.arena.get(unwrapped_spread).is_some_and(|node| {
-                            node.kind == syntax_kind_ext::CALL_EXPRESSION
-                                || node.kind == syntax_kind_ext::NEW_EXPRESSION
-                                || node.kind == syntax_kind_ext::TAGGED_TEMPLATE_EXPRESSION
-                        });
-                    let spread_request = if spread_is_call_like {
-                        base_request.contextual_opt(None)
-                    } else {
-                        base_request
-                    };
-                    let spread_type =
-                        self.get_type_of_node_with_request(spread_expr, &spread_request);
-                    let this_options_receiver_type = self
-                        .this_options_property_access_receiver(spread_expr)
-                        .map(|receiver_idx| self.get_type_of_node(receiver_idx));
-                    let is_contextual_this_options_any_spread = spread_type == TypeId::ANY
-                        && this_options_receiver_type
-                            .is_some_and(|receiver_type| receiver_type != TypeId::ANY);
-                    if !self.ctx.in_destructuring_target
-                        && spread_type == TypeId::ANY
-                        && !is_contextual_this_options_any_spread
-                    {
-                        has_any_spread = true;
-                    }
-                    // TS2698: Spread types may only be created from object types.
-                    // Only check in expression context — in destructuring targets,
-                    // `{ ...x }` is a rest binding (x receives remaining properties),
-                    // not a spread creation (reading x's properties), so the spread
-                    // validity check does not apply.
-                    // Also skip when TS2701 was already emitted (invalid rest target).
-                    let resolved_spread = self.resolve_type_for_property_access(spread_type);
-                    let resolved_spread = self.resolve_lazy_type(resolved_spread);
-                    let is_valid_spread = if self.ctx.in_destructuring_target || invalid_rest_target
-                    {
-                        true // rest binding in destructuring or already reported TS2701
-                    } else {
-                        crate::query_boundaries::type_computation::access::is_valid_spread_type(
-                            self.ctx.types,
-                            resolved_spread,
-                        )
-                    };
-                    if !is_valid_spread {
-                        self.report_spread_not_object_type(elem_idx);
-                    }
-
-                    // Short-circuit: when the object literal is a single spread
-                    // of a type parameter (e.g., `{ ...item }` where `item: T`),
-                    // preserve the type parameter as the result type. Expanding
-                    // to the constraint's properties would lose generic type
-                    // information, causing false TS2322 errors like
-                    // `Type '{ name: string }' is not assignable to type 'T'`.
-                    // Only when the spread is valid (no TS2698) — invalid spreads
-                    // like `T extends undefined` must not short-circuit.
-                    if is_valid_spread
-                        && obj.elements.nodes.len() == 1
-                        && properties.is_empty()
-                        && (crate::query_boundaries::common::type_param_info(
-                            self.ctx.types,
-                            spread_type,
-                        )
-                        .is_some()
-                            || crate::query_boundaries::common::contains_type_parameters(
-                                self.ctx.types,
-                                spread_type,
-                            ))
-                    {
-                        self.pop_object_literal_contexts(
-                            marker_this_type,
-                            partial_initializer_stack_index,
-                        );
-                        return spread_type;
-                    }
-
-                    // Check if the spread type is a union — if so, distribute
-                    // the spread over each union member: { ...A|B } → { ...A } | { ...B }
-                    let union_members_opt = crate::query_boundaries::common::union_members(
-                        self.ctx.types,
-                        resolved_spread,
-                    );
-
-                    // Guard against exponential blowup: if the cross-product
-                    // of branches would exceed a limit, skip distribution.
-                    let branch_count = if union_spread_branches.is_empty() {
-                        1
-                    } else {
-                        union_spread_branches.len()
-                    };
-                    let union_members_opt = union_members_opt.filter(|members| {
-                        // Only distribute when all members are object-like (not
-                        // false/null/undefined). Spreading primitives just
-                        // contributes {} which isn't useful to distribute.
-                        let all_object_like = members.iter().all(|m| {
-                            !self
-                                .ctx
-                                .types
-                                .collect_object_spread_properties(*m)
-                                .is_empty()
-                        });
-                        all_object_like && branch_count.saturating_mul(members.len()) <= 16
-                    });
-
-                    if let Some(members) = union_members_opt {
-                        // TS2783: Check if any earlier named properties will be
-                        // overwritten by required properties from this union spread.
-                        // A property triggers TS2783 when it is required (non-optional)
-                        // in ALL non-nullish members of the union.
-                        if self.ctx.strict_null_checks() {
-                            let non_nullish_members: Vec<TypeId> = members
-                                .iter()
-                                .copied()
-                                .filter(|m| !m.is_nullable())
-                                .collect();
-                            if !non_nullish_members.is_empty() {
-                                // Collect properties per member
-                                let all_member_props: Vec<Vec<_>> = non_nullish_members
-                                    .iter()
-                                    .map(|m| self.collect_object_spread_properties(*m))
-                                    .collect();
-                                // Find properties that are required in ALL members
-                                if let Some(first) = all_member_props.first() {
-                                    for prop in first {
-                                        if prop.optional {
-                                            continue;
-                                        }
-                                        let in_all =
-                                            all_member_props[1..].iter().all(|member_props| {
-                                                member_props
-                                                    .iter()
-                                                    .any(|p| p.name == prop.name && !p.optional)
-                                            });
-                                        if in_all
-                                            && let Some((prop_node, prop_name)) =
-                                                named_property_nodes.get(&prop.name)
-                                        {
-                                            let message = format_message(
-                                                    diagnostic_messages::IS_SPECIFIED_MORE_THAN_ONCE_SO_THIS_USAGE_WILL_BE_OVERWRITTEN,
-                                                    &[prop_name],
-                                                );
-                                            self.error_at_node(
-                                                    *prop_node,
-                                                    &message,
-                                                    diagnostic_codes::IS_SPECIFIED_MORE_THAN_ONCE_SO_THIS_USAGE_WILL_BE_OVERWRITTEN,
-                                                );
-                                        }
-                                    }
-                                }
-                            }
-                        }
-
-                        // Union spread distribution: fork current property set
-                        // into N branches, one per union member.
-                        has_union_spread = true;
-                        let mut new_branches: Vec<FxHashMap<Atom, PropertyInfo>> = Vec::new();
-
-                        // Collect properties from each union member for TS2783
-                        // and branching.
-                        let mut all_member_props: Vec<Vec<PropertyInfo>> = members
-                            .iter()
-                            .map(|m| self.collect_object_spread_properties(*m))
-                            .collect();
-                        remove_synthetic_missing_union_spread_props(&mut all_member_props);
-
-                        // TS2783: When a property is required (non-optional)
-                        // in ALL members of the union spread, it will always
-                        // overwrite any earlier named property.
-                        if self.ctx.strict_null_checks() && !named_property_nodes.is_empty() {
-                            // Find property names that are required in every member.
-                            let mut always_required: FxHashMap<Atom, bool> = FxHashMap::default();
-                            for (i, member_props) in all_member_props.iter().enumerate() {
-                                if i == 0 {
-                                    for prop in member_props {
-                                        always_required.insert(prop.name, !prop.optional);
-                                    }
-                                } else {
-                                    // Remove names not present in this member
-                                    always_required.retain(|name, required| {
-                                        if let Some(prop) =
-                                            member_props.iter().find(|p| p.name == *name)
-                                        {
-                                            if prop.optional {
-                                                *required = false;
-                                            }
-                                            true
-                                        } else {
-                                            false
-                                        }
-                                    });
-                                }
-                            }
-                            for (name, required) in &always_required {
-                                if *required
-                                    && let Some((prop_node, prop_name)) =
-                                        named_property_nodes.get(name)
-                                {
-                                    let message = format_message(
-                                            diagnostic_messages::IS_SPECIFIED_MORE_THAN_ONCE_SO_THIS_USAGE_WILL_BE_OVERWRITTEN,
-                                            &[prop_name],
-                                        );
-                                    self.error_at_node(
-                                            *prop_node,
-                                            &message,
-                                            diagnostic_codes::IS_SPECIFIED_MORE_THAN_ONCE_SO_THIS_USAGE_WILL_BE_OVERWRITTEN,
-                                        );
-                                }
-                            }
-                            // Clear named-property tracking for overwritten props
-                            for name in always_required.keys() {
-                                named_property_nodes.remove(name);
-                            }
-                        }
-
-                        let spread_order_base = spread_display_order_base;
-                        spread_display_order_base =
-                            spread_display_order_base.saturating_sub(SPREAD_DISPLAY_ORDER_STRIDE);
-                        for member_props in all_member_props {
-                            let member_props = rebase_spread_display_property_order(
-                                &member_props,
-                                spread_order_base,
-                            );
-                            if union_spread_branches.is_empty() {
-                                // First union spread: fork from the main properties
-                                let mut branch = properties.clone();
-                                for prop in member_props {
-                                    self.merge_spread_property(&mut branch, &prop);
-                                }
-                                new_branches.push(branch);
-                            } else {
-                                // Subsequent union spread: cross-product with existing branches
-                                for existing in &union_spread_branches {
-                                    let mut branch = existing.clone();
-                                    for prop in &member_props {
-                                        self.merge_spread_property(&mut branch, prop);
-                                    }
-                                    new_branches.push(branch);
-                                }
-                            }
-                        }
-                        union_spread_branches = new_branches;
-                        // Clear main properties so post-union properties
-                        // don't include pre-union ones when applied at the end
-                        properties.clear();
-                    } else {
-                        // When the spread type is/contains a type parameter,
-                        // track it for intersection creation at the end.
-                        // This preserves generic identity so that return types
-                        // of generic functions are properly instantiated at
-                        // call sites. Without this, spreading a type parameter
-                        // resolves to constraint properties, losing the generic
-                        // information and causing false TS2741/TS2322 errors.
-                        let is_generic_spread = is_valid_spread
-                            && (crate::query_boundaries::common::type_param_info(
-                                self.ctx.types,
-                                spread_type,
-                            )
-                            .is_some()
-                                || crate::query_boundaries::common::contains_type_parameters(
-                                    self.ctx.types,
-                                    spread_type,
-                                ));
-
-                        if is_generic_spread {
-                            generic_spread_types.push(spread_type);
-                        }
-
-                        let resolved_spread = self.resolve_lazy_type(spread_type);
-                        let resolved_spread = self.evaluate_type_with_env(resolved_spread);
-                        let resolved_spread =
-                            self.resolve_type_for_property_access(resolved_spread);
-                        let spread_props = self.collect_object_spread_properties(resolved_spread);
-                        // In thisless generic option patterns, `this.options.foo` can
-                        // temporarily resolve to `any` even though the containing call
-                        // gives this literal a concrete contextual target. Use that
-                        // target only for TS2783 overwrite diagnostics; keep type
-                        // construction based on the actual spread source.
-                        let spread_props_for_overwrite = if spread_props.is_empty()
-                            && spread_type == TypeId::ANY
-                            && is_contextual_this_options_any_spread
-                            && let Some(ctx_type) = contextual_type
-                        {
-                            self.collect_object_spread_properties(ctx_type)
-                        } else {
-                            spread_props.clone()
-                        };
-                        // Propagate index signatures from spread source.
-                        // When spreading an object with index signatures (e.g.,
-                        // `{ ...roindex }` where `roindex: { readonly [x: string]: number }`),
-                        // the result should inherit the index signatures (with readonly removed).
-                        // These are collected separately and only included in the final type
-                        // when the literal has no explicit (non-spread) properties, matching tsc.
-                        if (spread_props.is_empty()
-                            || !self.spread_source_is_unannotated_object_literal_binding(spread_expr))
-                            && !crate::query_boundaries::type_computation::core::is_fresh_literal_indexed_object(
-                                self.ctx.types,
-                                resolved_spread,
-                            )
-                        {
-                            use crate::query_boundaries::common::IndexSignatureResolver;
-                            let resolver = IndexSignatureResolver::new(self.ctx.types);
-                            let index_info = resolver.get_index_info(resolved_spread);
-                            if let Some(string_index) = index_info.string_index.or_else(|| {
-                                resolver.resolve_string_index(resolved_spread).map(|value_type| {
-                                    IndexSignature {
-                                        key_type: TypeId::STRING,
-                                        value_type,
-                                        readonly: false,
-                                        param_name: None,
-                                    }
-                                })
-                            }) {
-                                spread_string_index_signatures.push(string_index);
-                            }
-                            if let Some(number_index) = index_info.number_index.or_else(|| {
-                                resolver.resolve_number_index(resolved_spread).map(|value_type| {
-                                    IndexSignature {
-                                        key_type: TypeId::NUMBER,
-                                        value_type,
-                                        readonly: false,
-                                        param_name: None,
-                                    }
-                                })
-                            }) {
-                                spread_number_index_signatures.push(number_index);
-                            }
-                        }
-
-                        // TS2783: Check if any earlier named properties will be
-                        // overwritten by required properties from this spread.
-                        // Only when strict null checks are enabled.
-                        // TSC checks constraint properties even for generic spreads,
-                        // so we do too (unlike type construction, approximations are fine here).
-                        if self.ctx.strict_null_checks() {
-                            for sp in &spread_props_for_overwrite {
-                                if !sp.optional
-                                    && let Some((prop_node, prop_name)) =
-                                        named_property_nodes.get(&sp.name)
-                                {
-                                    let message = format_message(
-                                        diagnostic_messages::IS_SPECIFIED_MORE_THAN_ONCE_SO_THIS_USAGE_WILL_BE_OVERWRITTEN,
-                                        &[prop_name],
-                                    );
-                                    self.error_at_node(
-                                        *prop_node,
-                                        &message,
-                                        diagnostic_codes::IS_SPECIFIED_MORE_THAN_ONCE_SO_THIS_USAGE_WILL_BE_OVERWRITTEN,
-                                    );
-                                }
-                            }
-                        }
-
-                        // After TS2783 check, clear the named-property tracking
-                        // for properties that the spread overwrites (so only the
-                        // first occurrence can trigger the diagnostic, not later
-                        // spreads which are spread-vs-spread and exempt).
-                        for prop in &spread_props_for_overwrite {
-                            named_property_nodes.remove(&prop.name);
-                        }
-
-                        let spread_props_for_display = rebase_spread_display_property_order(
-                            &spread_props,
-                            spread_display_order_base,
-                        );
-                        spread_display_order_base =
-                            spread_display_order_base.saturating_sub(SPREAD_DISPLAY_ORDER_STRIDE);
-                        for prop in &spread_props_for_display {
-                            self.merge_spread_property(&mut properties, prop);
-                        }
-
-                        // Also apply non-union spread to any existing union branches
-                        for branch in &mut union_spread_branches {
-                            for prop in &spread_props_for_display {
-                                self.merge_spread_property(branch, prop);
-                            }
-                        }
-                    }
+                if let Some(spread_type) = self.process_object_literal_spread_element(
+                    elem_idx,
+                    obj.elements.nodes.len(),
+                    &base_request,
+                    contextual_type,
+                    marker_this_type,
+                    partial_initializer_stack_index,
+                    &mut properties,
+                    &mut named_property_nodes,
+                    &mut union_spread_branches,
+                    &mut spread_string_index_signatures,
+                    &mut spread_number_index_signatures,
+                    &mut generic_spread_types,
+                    &mut has_spread,
+                    &mut has_any_spread,
+                    &mut has_union_spread,
+                    &mut spread_display_order_base,
+                ) {
+                    return spread_type;
                 }
             }
+
             // Other element types (e.g., unknown AST node kinds) are silently skipped
         }
